+++ apps/web/src/app/main/community/[tab]/[postid]/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import { Eye, MoreVertical } from 'lucide-react';
+import { Eye } from 'lucide-react';
 import type { GetPostResponse } from '@/api/posts';
 import LikeButton from '@/app/main/community/[tab]/[postid]/components/LikeButton';
 import ImageSlider from '@/components/ImageSlider';
 import PostProfile from './components/PostProfile';
-import CommentProfile from './components/CommentProfile';
+import CommentList from './components/CommentList';
 
 // 게시글 더미
 const dummyPost: GetPostResponse = {
@@ -29,48 +29,6 @@ const dummyPost: GetPostResponse = {
     userType: 'resident',
   },
 };
-
-// 댓글 더미
-type CommentUser = {
-  nickname: string;
-  profileImageUrl: string;
-  userType: 'founder' | 'resident';
-};
-
-type Comment = {
-  id: number;
-  user: CommentUser;
-  content: string;
-  likeCount: number;
-  edited?: boolean;
-  createdAt: string;
-};
-
-const dummyComments: Comment[] = [
-  {
-    id: 11,
-    user: {
-      nickname: '민수',
-      profileImageUrl: '/somoon/default_somoon.svg',
-      userType: 'resident',
-    },
-    content: '오 여기 가봤는데 줄 길더라구요. 꿀팁 감사!',
-    likeCount: 3,
-    createdAt: '2025-08-07T03:00:00Z',
-  },
-  {
-    id: 12,
-    user: {
-      nickname: '앨리스',
-      profileImageUrl: '/somoon/default_somoon.svg',
-      userType: 'founder',
-    },
-    content: '사진 보니 또 가고 싶네요 :)',
-    likeCount: 1,
-    edited: true,
-    createdAt: '2025-08-07T05:40:00Z',
-  },
-];
 
 export default function PostPage() {
   const post = dummyPost;
@@ -122,22 +80,7 @@ export default function PostPage() {
         </div>
 
         {/* 댓글 리스트 */}
-        <section className="pt-6 space-y-4">
-          {dummyComments.map((comment) => (
-            <CommentProfile
-              key={comment.id}
-              nickname={comment.user.nickname}
-              profileImageUrl={comment.user.profileImageUrl}
-              userType={comment.user.userType}
-              likeCount={comment.likeCount}
-              edited={comment.edited}
-              createdAt={comment.createdAt}
-              action={<MoreVertical className="w-4 h-4" />}
-            >
-              {comment.content}
-            </CommentProfile>
-          ))}
-        </section>
+        <CommentList />
       </main>
     </div>
   );
